Manually merged with Ryan and Troy's code
