+++ src/SimpleChatServer/EchoServer1.java
@@ -1,6 +1,7 @@
 package SimpleChatServer;
 
 import java.io.*;
+import java.util.ArrayList;
 import java.util.HashMap;
 
 import client.ChatClient1;
@@ -27,11 +28,12 @@ public class EchoServer1 extends AbstractServer
   //Class variables *************************************************
 	private boolean closed; 
 	private ChatIF myServerUI;
-	private HashMap<String, String> accounts;//added by Shouheng
   /**
    * The default port to listen on.
    */
   final public static int DEFAULT_PORT = 5555;
+    private ArrayList<Channel> channels;
+    private HashMap<String, String> accounts;//added by Shouheng
   
 
   //Constructors ****************************************************
@@ -46,8 +48,11 @@ public class EchoServer1 extends AbstractServer
     super(port);
     myServerUI = serverUI;
     closed = false;
-    accounts = new HashMap<String, String>();
-    accounts.put("guest", "123");
+    
+    accounts = new HashMap<String, String>();//added by Shouheng
+    accounts.put("guest", "123");//added by Shouheng.
+    
+    initializeChannels();
     try {
     	listen();
     }
@@ -58,6 +63,10 @@ public class EchoServer1 extends AbstractServer
   
   public ChatIF serverUI() {
   	return myServerUI;
+  }
+  
+  private void initializeChannels() { 
+	  channels = new ArrayList<Channel>();
   }
 
   //Instance methods ************************************************
@@ -79,9 +88,70 @@ public class EchoServer1 extends AbstractServer
 	 
   }
   
-  public void handleMessageFromUser(String message){
+  public void addChannel(Channel chl) {
+	  channels.add(chl);
+  }
+  
+  
+  private void sendToChannelClients(Object msg, String channel) {
+	  for (int j = 0; j < channels.size(); j++) {
+		  Channel chl = channels.get(j);
+		  if (chl.getChannelName().equals(channel)) {
+			  Object[] channelClients = chl.enumerateClients();
+			  for (int i=0; i<channelClients.length; i++)
+			    {
+			      try
+			      {
+			        ((ConnectionToClient)channelClients[i]).sendToClient("Server MSG "+ chl.getChannelName()+"> "+msg);
+			      }
+			      catch (Exception ex) {
+			    	  serverUI().display("Error in sending message");
+			      }
+			    }
+			  return;
+		  }
+		  else {
+			  serverUI().display("Channel with the name " + channel + " does not exist.");
+		  	  return;
+		  }
+	  }
 	
-	  if(message.charAt(0) != '#')
+  }
+  
+//Written by Shouheng Wu
+  //This method checks whether the given ID is already an existing user
+  public boolean checkExistingAccount (String id){
+	  if (accounts.containsKey(id)){
+		  return true;
+	  }
+	  else{
+		  return false;
+	  }
+  }//end checkExistingAccount
+  
+  //Written by Shouheng Wu
+  //This method creates a user account by adding a id/password combination to the hashmap accounts
+  public void setNewAccount(String id, String password){
+	  accounts.put(id, password);
+  }//end class
+  
+  //Written by Shouheng Wu
+  //This method returns true if the provided password is correct
+  public boolean checkPassword(String id, String password){
+	  if(accounts.get(id).equals(password)){
+		  return true;
+	  }
+	  else{
+		  return false;
+	  }
+	  
+  }//end checkPassword
+  
+  public void handleMessageFromUser(String message){
+	  if (message.charAt(0) == '@') {
+		  sendToChannel(message.substring(1));
+	  }
+	  else if(message.charAt(0) != '#')
 	    {
 		  sendToAllClients("SERVER MSG>" + message);
 	    }
@@ -93,6 +163,21 @@ public class EchoServer1 extends AbstractServer
 	  
 	  
   }//end handleMessageFromUser
+  
+  public void sendToChannel(String message) {
+	  String channelName;
+	  int indexBlank = message.indexOf(' ');
+	  if(indexBlank == -1) {
+		  channelName = message;
+		  String msg = "";
+		  sendToChannelClients(channelName, msg);
+	  }
+	  else {
+		  channelName = message.substring(0, indexBlank);
+		  String msg = message.substring(indexBlank + 1);
+		  sendToChannelClients(msg, channelName);
+	  }
+  }
   
   public void createAndDoCommand(String message){
 	  String commandStr; 
@@ -164,35 +249,6 @@ public class EchoServer1 extends AbstractServer
   protected void setClosed(Boolean status){
 	  closed = status;
   }//end setStatus
-  
-  //Written by Shouheng Wu
-  //This method checks whether the given ID is already an existing user
-  public boolean checkExistingAccount (String id){
-	  if (accounts.containsKey(id)){
-		  return true;
-	  }
-	  else{
-		  return false;
-	  }
-  }//end checkExistingAccount
-  
-  //Written by Shouheng Wu
-  //This method creates a user account by adding a id/password combination to the hashmap accounts
-  public void setNewAccount(String id, String password){
-	  accounts.put(id, password);
-  }//end class
-  
-  //Written by Shouheng Wu
-  //This method returns true if the provided password is correct
-  public boolean checkPassword(String id, String password){
-	  if(accounts.get(id).equals(password)){
-		  return true;
-	  }
-	  else{
-		  return false;
-	  }
-	  
-  }//end checkPassword
   
   protected synchronized void clientException(ConnectionToClient client, Throwable exception){
 	  sendToAllClients("SERVER MSG>  " + client.getInfo("id") + " has disconnected.");
